Code cleanup and organization
This way SHIELD can see the shield-crypt error if something happens.
